Introduce DB migration 9 & fix a bug introduced at 1.27 with new DB
- In 1.27.0 a bug was introduced where new DBs that were created (or
any DBs that needed no migration ran -- but that is only new DBs) did
not write anything in the DB for last_migration_version. This means
that even if there came a new version, like from this commit, they
would not actually apply it. Fixed it so that even if no migrations
run the last data migration version of the current rotki version is
written in the DB

- Introducing migration version 9. This is to address
https://github.com/rotki/rotki/issues/5560. The issue here was that
all history events for on-chain transactions have location:
Blockchain. Any they contain no way to identify the chain. With this
migration we add a location for Ethereum and a location for
Optimism. And then we check all history events, check the
corresponding transaction hash and the chain it is in and adjust the
location accordingly. Having just general location blockchain was a
useless field in the past.