Rework get_ignored_assets & asset search ignores ignored tokens now
Made some changes so that all asset searches api endpoints ignore all
ignored assets during search.

Also removed the old `get_ignored_assets` function as it was really
heavy-handed.

It used to get all ignored asset ids, turn them into Assets (so for
each ignored asset, that is a roundtrip to the global DB) and then if
that does not work even delete it from the user DB.

Replaced it with a simple function fetching the identifiers of the
ignored assets and its usage with the new function everywhere.

Perhaps it would make sense to also have either a periodic query, or
an one-off function at DB start to do the removal of unknown assets
from the ignored assets. But in another commit/PR.