Update built-in globalDB with migration1 ran and ilk cache pulled
To make sure these don't need to be pulled all the time, we update the
built-in globalDB with the state it should have after migration 1 runs
and the cache has been pulled. Adds 59 ilk types and their respective
contracts and metadata.

Not all of them have join adapters that conform to the GemJoin interface

We achieve this query by starting without a global.db and with a clean
account. At the end of unlock we add the query_ilk_registry function
and then sys.exit()