Fix maybe_kill_running_tx_query_tasks edge case
Fixed an edge case where maybe_kill_running_tx_query_tasks ran 2 times
between two task schedulings. The second time the greenlet was already
killed but had not been removed by the list and was still
there. Trying to access kwargs of a dead greenlet raised an exception.

Fixed it and also wrote a regression test for it.