Claiming COMP after comptroller refill correct decoding
This fixes a bug where there was COMP reward accumulated in the
controll but failed to be paid out in an older transaction due
to (probably) being out of COMP.

Code is here:
https://etherscan.io/address/0xBafE01ff935C7305907c33BF824352eE5979B526#code

After being refilled the Distributed[Supplier/Borrower] events won't
match the transfers.

This commit fixes that bug by forcing the transfer search even if
compdelta is zero, as accumulation of rewards can have happend in a
previous transaction.