Dont query all block internal transactions, but all parent hash ones
This can save **a lot** of etherscan queries for when decoding
transactions of a block with many internal transactions.

Before, in order to find internal transactions of a transaction we
used to query for the same block's internal transactions.

That added a lot of extra queries, as a block can have many many
transactions, each with their own internal transactions, and all were
pulled, since an internal transaction's parent is also pulled when the
internal is pulled.

This is easily solved since etherscan has recently added a query for internal
transactions of a transaction hash.