If known protocol token chain price query fails, fallback to oracles
The code before exited immediately after the on-chain query failed,
thinking that the external oracles will not be able to give an answer.

But that's misguided. Both coingecko and defillama have lots of those
special tokens and even the ability to query by address (which we
already utilize at least for defillama)