Configure web3.py middlewares
We found that web3.py was making one extra rpc call after every eth_call. After investigating
we discovered that it was the on-chain validation in the web3.py middleware. Since we don't need
it for our usage atm we decided to remove this middleware and save rpc calls.

We have also removed middlewares that we don't need like gas estimation and ens resolution