Protect with a lock contract query in packaged db
It could happen that two different requests tried to access at the same time the packaged db
to copy a contract and then the task would fail because the packaged db would be already attached
by one of them.

This was easily seen while visiting a page with multiple ens avatars when the contract was still not copied.