Fix packaged_db is locked error
Unfortunately the previous approach here:
https://github.com/rotki/rotki/pull/6058 was not tested and as such
was not a solution.

The problem happens only at detaching from the packaged DB and only
when there is an active transaction for the same connection.

https://stackoverflow.com/questions/56243770/sqlite3-detach-database-produces-database-is-locked-error

So the approach I followed was to guard with the transaction lock when
we try to detach a DB.

Also reproduced in a test.