"Fix" for 5432
This is a "temporary fix" for the error binding parameter 0 flaky
error as seen in https://github.com/rotki/rotki/issues/5432.

If I understand it correctly the bug is in the python sqlite drivers
and as such this is just a patch until it's solved properly.

We have a test to reproduce so we should periodically check as we
upgrade python versions to see if this is fixed and then we can revert.