Remove query of size for past PnL reports
A user reported to us that the app was freezing for him after loading a huge pnl report.
By reading the logs it took around 4 mins for the endpoint to return a value. The reasons for
the issue where:

1. The reports endpoint is called sync and not async
2. It was performing a heavy sql query fetching the size of the reports

This was blocking the backend from receiving any other request and also making as if no report
was getting generated since it was not being displayed on the frontend.

We discussed in discord and didn't see any meaningful reason to have it so we are removing the size
on disk info for now.