Add a custom configurable requests session for tests
This session provides timeout by default and also
a 10x max connection pool size. It is also further customizable via
pytest fixtures.

Added it only a few tests in order to avoid hitting the normal
requests connection blocking timeout that we seem to see in the CI.

Also in the tests that this session replaced pure requests