Downgrade urllib3 to 1.26.14
The reason for this is that from 1.26.15 and on seem to cause our CI tests
to block/hang in an unpredictable way and we are not sure exactly why.

From the diff betweem 1.26.14 and 1.26.15 the most relevant code seems
to be
https://github.com/urllib3/urllib3/commit/a7ce8e0881c94800b14687145ee11940246d2b22
which addresses https://github.com/urllib3/urllib3/issues/2645.

But still not 100% sure if this is the cause or why that would be.