For timestamp_to_date use the timezone of the timestamp
Before it was always using the now() timezone. Which would be correct in most
cases except for cases when DST or similar is activated.

So it would show CEST all year round for people in Germany. Now it
will show CEST with DST on and CET when DST is off.