Fix edge case in uniswap liquidity withdrawal
There are two types of withdrawals in uniswap LPs, one where eth is withdrawn and another
where WETH is withdrawn. As it was before the WETH withdrawal was not properly decoded since
it was looking for ETH instead of ETH due to a condition in the assert searched.

The solution proposed is to look for eth and weth in the events and once found create the correct
transform instruction