Fix kyber decoder for trades
The decoder was using the wrong tx log topic to filter trandes and it was failing to decode the
transaction because it matched againt an event that didn't have the expected fields.

This was not noticed in tests because I didn't use all the logs and just some of them. By mocking
the two tests it is now properly tested.

Also fixed an issue with uniswap front running the kyber decoder