Make the total lost amount always positive
After 9189ce7 the aave eth events got correctly decoded
and this uncovered an issue in the aave stats.In the test the total lost amount was getting
displayed as negative instead of positive.

It didn't appear before because it came from a liquidation event that was being correctly accounted
and the debt event that existed before was getting cancelled.