If block mev reward already exists dont fail and delete evm receive
This will happen if the transaction with the receive ETH event of the
mev reward is redecoded. Or when all evm transactions are reset. The
event receive is seen again and when attempted to be combined it
failed due to the combination having already happened.

So we now added code to handle this