Querying asset existence with case insensitive ID now normalizes it
When resolving an asset via `resolve()` the returned resolved asset, has the
normalized identifier. But when checking for asset existence with
`check_existence()` the asset can be seen to be in the DB but its
identifier will not be normalized.

This can result to all the following assets moving around in our
codebase: Asset('ETH'), Asset('eth'), Asset('eTh'), Asset ('etH')
etc. While the normalized identifier is 'ETH'.

This hit us recently in the icons query, where the frontend was
sending us 'eth' instead of 'ETH', and this was seen as a valid asset,
but its icon filepath was not found in Linux since it expected
"$ICON_DIR/eth_small.png" but did not find it. What's interesting is
that this works fine for Mac and Windows since filepaths there are
case insensitive.