Ensure that arbitrum nodes get populated
It can happen for a new user that the arbitrum nodes don't get copied in the globaldb
and then the user db doesn't get them. For polygon we released an update in the remote data
repo with the endpoints but if you had an old globaldb and create a new user before the task
that queries the remote information for rpc nodes is scheduled then the globaldb doesn't contain
any arb rpc node and then the user db doesn't copy any on load. They would get copied once the task
triggers.