Account for MEV rewards in the accounting process
A user reported in discord that MEV rewards didn't appear correctly in the CSV summary.
What I found is that we assumed in the code that there was a evm event with a receive of the MEV
from the builder but in ae358d2956ed121ae53eb54580fefc48279d41be we changed this and now events are combined.

What I did was:
1. Make sure that events are combined before running the PnL report
2. Allow to process evm events normally