Fix DB plaintext locked -- third try
Fix https://github.com/rotki/rotki/issues/5038 ... for the third time.

The idea is to put the entire export unencrypted DB under a critical
DB section.

It's needed because a context switch
from inside this execute script can result in:
1. coming into this code again from another greenlet which can result
to DB plaintext already in use
2. Having a DB transaction open between the attach and detach and not
closed when we detach which will result in DB plaintext locked.