Fix missing gitcoin donations in optimism
Taking this huge transaction as an example it was noticed that
multiple donations were missing:
https://optimistic.etherscan.io/tx/0x5d85b436f5f177de6019baa9ecebae285e0def4924546307fac40556bece4cd7

This was due to not having many voting_impl addresses in the
decoder. This is unfortunately a manual process which will need fixing
often.

Also fixed a problem in the decoder where the original big transfer of
the token to the contract for splitting was not removed.