Adjust fixtures and fix various tests
By debugging some VCR tests I realized our mocks that build up
rotkehlchen instance for tests are flawed in multiple places.

This is not a full refactor and we really should move away from so
many mocks.

This is simply trying to fix parts of mocking that were wrong,
specifically about enforcing nodes to connect to (they were not --
default rpc nodes were used instead) and to actually connect to them
if needed.

Adjusting tests and other fixtures as needed