Fix an old circular import with some ethereum/modules and some tests
This is a quite old circular import that has been hitting us whenever
something in the import line was passing through
ethereum/modules/init.py and invoked all the various imports which
produced a knock-on effect that ended up circling the imports.

The solution is to use something introduced long time ago to fix
this. A mapping that is loaded dynamically whenever an ethereum module
needs to be loaded.