Fix resolve_asset_to_class. If type in DB is not what we ask bail.
And by bailing earlier we don't end up overwriting the user's global
DB with the packaged global DB.

What was happening was that at some point all assets were attempted to
be resolved to FIAT assets. Even if they were not FIAT assets. This
basically got us in this code for each asset, and if the asset was in
the constant assets it automatically replaced itself with what was in
packaged DB.

Now this will no longer happen. Only if asset is not in the global DB
or is not what was expected and in the packaged DB it is, will the
replacement happen.