When querying current prices use asset resolution properly
Before the code was doing asset resolution as an attempt to try and
see if it's a certain type of asset. For example for FiatAsset, or
EvmToken and used some kind of duck typing with exception suppression.

That's super ugly and hid a bug that made us overwrite user's
globalDB. (see previous commit)

Now instead of doing that we just resolve the asset once and then act
according to type.