Remove session_globaldb fixture and dependencies
That fixture was not useful and was in fact broken. It became obvious
when I tried to use it. We can't have a different globalDB fixture for
session and a different for function scope in tests.

The reason being it's a singleton and the function one will modify the
session one. So things get mixed up pretty badly causing undefined
behaviour. Which can hit you when running tests with "trying to
operate on a closed DB" if the cleanup of the function scope globalDB
is done before a new test and a session scope fixture needing session
globalDB tries to read it.

Generally bad idea