Handle settings keys that are not present in the db for accounting rules
It might be possible that an accounting setting is not present in the db when it takes
the default value. This commit ensures that when a setting is used in the rules then it
is present in the database.

Also fixes an issue with the filters not being applied to the linked rules when filtering