Add decoder for migration of MATIC to POL in ethereum mainnet
Fix edge case of maybe_reshuffle_events

If another event was in between two of the events we wanted to
preserver order of then there could be a case where previous code
failed.

Example:

spend: 351, approve: 352, receive: 353

This would not put spend and receive one after the other. Now it will.