Adjust event category details mapping and move direction
EventCategory is now a more complicated enum, which also has the
direction in it.

The event_category_details which is returned to the frontend for
displaying icons and more per event type is now configurable also by
counterparty on top of type/subtype.

This is now only used for spend/fee and counterparty gas as a
specialization of spend/fee.