Fix PnL report for case when non-history Event follows a swap
What was happening is that the code assumed any event from the
iterator was a HistoryEvent. So if during the processing of a swap we
peeked ahead to see if there is a fee and the next event was anythign
else, such as ValidatorDailystats then BOOM.

Ignoring mypy typing with a typecast was a bad mistake. Mypy had
caught this error.

Fix #7009