Move coinlist to DB cache and abstract out for oracles that use it
This essentially removes the cryptocompare_coinlist.json file cache
that was in the data directory. And also deletes it at data dir
upgrade.

Also removes the coingecko coinlist in memory cache.

Finally created some fixtures to stop tests that need to query those coinlists
(tests that add/edit/deal with assets in the api) from doing
so. Instead those fixtures create a cache beforehand so the queries
are no longer needed.

Signed-off-by: Lefteris Karapetsas <lefteris@refu.co>