Fix deletion of staking daily stats at validator refresh
During the DB upgrade we were adding new fields to the eth2_validators
table and making the validator index being a UNIQUE key instead of a
primary key.

When detecting new data for validators the code was doing INSERT OR
REPLACE which essentially replaced the entire row with a new one.

The problem here is that daily stats is tied to eth2 validators via
foreign key connection to the validator index.

This replacing dropped all eth daily stats. We now do a more
complicated query to detect if it's a new validator addition in which
case we have to insert it or if it's an old one in which case we have
to update some keys. Then we need to find which keys to update.

Signed-off-by: Lefteris Karapetsas <lefteris@refu.co>