Fix duplicate entries in globaldb conflicts
When updating assets it can happen that two remote updates by mistake
add the same token and maybe with extra data. Before this change the
multiple conflicts were being returned causing errors in the logic
for the api consumer since it was assumed uniqness.

This PR ensures that only the latest appearence is took as conflict