Fix Eigenlayer pending withdrawal balances
If the timing of the eigenlayer pending withdrawal completion came
before the saving of the start of pending withdrawal was in the DB,
then the pending withdrawal was never marked as complete.

This resulted in those cases of completed withdrawals counting again
in the balances.

This commit fixes this by forcing redecoding of completion if a
completed withdrawal is found without having a match to a pending
withdrawal whenever eigenlayer balances are queried.

Signed-off-by: Lefteris Karapetsas <lefteris@refu.co>