Fix addressbook
Set the blockchain value to a constant string instead of None
since the primary key of the table includes the nullable column.

This avoid having duplicated entries. Also ship a data migration
for both the user db and the globaldb to update the tuples.
In the case of duplicates we delete them except the one with the highest
row id.