Temporary fix for Eigenlayer validators exit/exiting
This is the bugfixes part of https://github.com/rotki/rotki/issues/8397.

Essentially we no longer rely on history events processing to detect a
validator has exited. That is due to the fact we may track a validator
but not its exit event due to not tracking the withdrawal address
directly. This happens for example in Eigenlayer native restaking.

This temporary fix is a hack considering that if "enough time" has
passed since the exit was started (the time of withdrawable_ts) and
now then we consider it exited.

Long term solution will happen in develop/1.35.0 as per the issue description

Signed-off-by: Lefteris Karapetsas <lefteris@refu.co>