Introduce a has_premium_check function
This is just to combine the premium is not None check with the active
premium check. In theory this way if a key expires and after few hours
the `Premium` object detects it in all the places where this is
checked they will switch to non premium.

While before this change even if you have a `Premium` object with
expired keys and `status = NOT_ACTIVE` it would count them as premium
being active

Signed-off-by: Lefteris Karapetsas <lefteris@refu.co>