Add untested gnosispay logic for refunds
I am making some assumptions here.

1. There is no onchain logs for these. It's always a transfer from the
spender to the user's gnosis pay safe.
2. Not sure of how the final api response would look like for a refund. At the moment
in my data I have a refund with same tx_hash as payment but this is
probably inconmplete as the refund has not yet happened. I assume
there will be another transaction hash in the transactions entry.

Signed-off-by: Lefteris Karapetsas <lefteris@refu.co>