Use new balance scanner contract
This commit introduces a new balance scanner contracts.
It is writting in vyper and is important because in case of a revert
calling the balanceof for a contract it returns 0 instead of failing
the transaction.

This change solves an issue where a bad token could make rotki not query
balances properly.

Regarding the chunk size I've checked different sizes

```
chunk size, time, node, seed, comments

375, 5.452023983001709,  flashbots
375, 5.356801986694336,  flashbots , 42
375, 8.883646965026855,  cloudflare
375, 10.064039945602417, cloudflare
375, 6.45735502243042,   cloudflare
375, 5.277329921722412,  cloudflare, 80
375, 4.732897043228149,  ankr      , 90
500, 4.8045032024383545, ankr + flashbots, 90,  ankr and cloudlare had a single call out of gas
480, 5.05453896522522, ankr, 90, ankr had a single call out of gas. Cloudflare okey
460, 3.771683931350708, ankr + cloudflare, 90, only 1 request to ankr failed
460, 4.571718215942383, ankr + cloudflare, 90, only 1 request to ankr failed
460, 6.436861991882324, flashbots, 42
460, 3.726022243499756, flahbots + ankr, 80
460, 7.433700084686279, cloudflare + ankr, 460
```

and the optimal was 460. I'm under the impression that a token consumes a little bit more of gas and
makes the transaction fail.