Fix wrong range queried in gnosis logs
The problems is that the range queried for logs was always of the form [deployed block, latest]

The reason is that the function is called in https://github.com/rotki/rotki/blob/87dc0db0a46e67c1497ff0ed0abbab4697352348/rotkehlchen/chain/evm/transactions.py#L161
with the data that comes from the api but the frontend never sends a start/end timestamps
so the range was in fact always [0, now].

Then the logic was

```python
        with self.database.conn.read_ctx() as cursor:
            cursor.execute('SELECT MAX(start_ts), MIN(end_ts) FROM used_query_ranges WHERE name LIKE ?', (f'{BRIDGE_QUERIED_ADDRESS_PREFIX}%',))  # noqa: E501
            result = cursor.fetchone()
            if result is not None:
                max_start, min_end = result

        if max_start is not None:
            if max_start <= from_ts and min_end >= to_ts:  # type: ignore # min_end not None
                return  # no need to do anything

            # else, adjust query
            from_ts = min(max_start, from_ts)
            to_ts = max(min_end, to_ts)  # type: ignore # min_end is not None
```

but in the db the ranges are stored in the form [0, end_ts] so

```
from_ts = min(max_start, from_ts)
```

was always `min(0, 0)` and to_ts was always now.

How I noticed this was by a change in the logs in my last PR that states the ranges saved to the database
and I could see that it was starting from the deployed block always.

Why didn't we see it in the tests since we have a check to ensure that the info wasn't queried again?
The test was calling

```
gnosis_transactions.get_chain_specific_multiaddress_data(
            addresses=gnosis_accounts,
            from_ts=Timestamp(0),
            to_ts=now,
        )
```

and the logic evaluated

```
        if max_start is not None:
            if max_start <= from_ts and min_end >= to_ts:  # type: ignore # min_end not None
                return  #
```

where `max_start <= from_ts and min_end >= to_ts` is `0 <= 0 and now >= now` and therefore no extra call was performed.

This PR changes the logic to:

1. Ignore the from and to. It introduced a bug because is not what the developer expected and we use the query ranges.
2. Properly query the widest range for logs
3. Ensure that if new addresses are added then we query logs for them