Limit arbiscan token queries
So when the balances are queried we do two things.

1. Query base token balances (ETH, xdai..)
2. Query evm tokens

`1.` always uses the default call order. that is a randomized order and etherscan can be the node queried.
`2.` For 2 we do something special. Since we split the tokens in chunks we have a funciton `get_chunk_size_call_order`. https://github.com/rotki/rotki/blob/da76cb8e347543bb2f539a6731c8666bc382eacb/rotkehlchen/chain/evm/tokens.py#L110 (screenshot)

We changed the way nodes are connected and if in the step `1.` we query eth balances using etherscan then there is no node connected. THen the check `connected_to_any_web3` becomes False and the token balance query is made using etherscan.

Now the second issue comes and it is that arbiscan has a lower limit in the amount of tokens that can be queried and that is not hanlded in that logic