Fix cc price queries for special assets
In a refactor the return type of a function was modified and
it broke the assumtpion about types in other part of the code.

It went unnoticed since we used getattr. I've changed it to a case
so we can see by searching for the function name where it is used and help
us in future refactors. Also added a test