Address connection errors
When the app is left running for a long time and I come back to it I see
connectionError: connection clossed by remote server. This happens to me mostly
with kraken.

After some investigation the cause seems to be the keep alive property of the connection.
The session will use the connection that has expired and get this error from the server.

It is not a behaviour from requests but internal to urllib. The solution seems to be to
use the httpAdapter class and enforce a retry, by default it doesn't happen. From the
rquests docs:

https://docs.python-requests.org/en/latest/api/#requests.adapters.HTTPAdapter

```
max_retries – The maximum number of retries each connection should attempt.
Note, this applies only to failed DNS lookups, socket connections and
connection timeouts, never to requests where data has made it to the server.
By default, Requests does not retry failed connections.
```