Change the order querying bybit
Bybit allows to query at most 2 years of trades and you can query at most 7 days
per request. We were using a sliding windows and querying from the closest timestamp to now
towards the past. In slow scenarios this affected the actual time that we can query since the oldest
timestamp after 7 minutes of queries is what we set initially + 7 mins and we weren't having into
consideration this delay. There was a safety net of 5 mins but for some users it was exceeded.

To fix the issue now the query starts by querying the oldest possible timestamp first and goes
towards the future. This prevents this issue with the time window moving.