Do not double count Aura positions
There are some pretty glaring errors in the logic here.

1. There was an AuraFinancebalances module. Which seems completely
unnecessary. As it just does a balanceOf all aura vault tokens that
the user has. That's already done anyway through the normal token
balance query.
2. There was a logic for BalancerV2 and v1 which seems to be not
needed. Because it asks the BalancerAPI and the balancer API tells you
that you have "staked" tokens and "total" tokens. We take the
total. The problem here is that staked means staked in Aura. So
remember thet we already double counted Aura? Now we are triple
counting it. I think we would also end up double counting the
non-staked ones and the staked in balancer since all of them would end
up giving back a token whose balanceOf would count the balance
again.