Fix gnosispay data redecoding due to API bug
So our logic is trying to redecode gnosis pay transactions by asking
for the gnosis pay data -10 seconds and +10 seconds from the
transaction hash.

For a very limited rare case of transactions that returned an empty
list.

Which is a bug in the API of gnosis pay.

By increasing the range to -30 / + 30 it seems to return the
transactions for the cases we tested that were missing. This also
needs to be reported to gnosis pay API.

Also adding few more logs.