Do backups in a new thread
Closes #9772

The backup logic compresses and encrypts the database before uploading it.
This is a very intensive CPU logic and was blocking the app since it doesn't
context switch. This commit spawns it in a thread allowing the api to continue
answering requests.

It increases the threadpool max size to allow this additional thread for
backups.