Sanitizing Rotki Backend Start-Up Arguments in Docker Entrypoint
Python's `subprocess.Popen` accepts a list of strings and PathLikes
as arguments to start a process. The `entrypoint.py` script reads
environment variables and settings from a config file, casting
numeric-expected values to `int` as a first line of defense against
mis-configuration. The casted values are added to the arguments
list for `subprocess.Popen`. When such values are configured, the
startup fails because of an exception from `Popen`.

This fix introduces a quick sanitize of the arguments list, just
for the use with `subprocess.Popen`, keeping the defense line
intact.

For a later refactoring, a proper value validation should be
considered. This could provide better / more helpful error messages
in case of a misconfiguration.