Improve performance of balance query
When we query protocol balances the filter was heavy and could be optimized.
The main problem is that our filters don't allow to query pairs of [event_type, event_subtype]
and we have a filter in python to discard the combinations that we don't need.

To improve the performance of this query the filter now includes all the [event_subtype] but not
the combinations. This allows to improve the performance of this section since the for loop
needs to process less events later.

Also I've adjusted the joins that are made when querying using EvmFilterQuery. This didn't affect
in a clear manner but I believe is a good thing to have