Fix maybe_upload_data_to_server deadlock
This commit does two things:

1. Creates a test for the deadlock seen ocasionally when
maybe_upload_data_to_server and redecoding happened at the same
time.
2. Fixes it by making sure the new thread only does the compressing
and encrypting but not the DB reading.

It's a more generic deadlock where due to the the fact that the
threadpool creates a new greenlet in another thread and accesses the
database in that thread and perhaps context switches out of it stuff
completely lock.

The tests kind of attempts to emulate this. And ~50% of the runs of
that test with the code before the commit failed with database locked
error messages. While with the change it should always pass.