New internaltx and multiconsolidation requests
- Added gas and gas_used to the internal tranasaction table. The
reason for this change is that the current implementation has a
primary key that's not unique for cases like this
https://etherscan.io/tx/0x72f5e619a8f521d874652ec5c09ea22e329ed3a990012e1fe6548d5a07e1959c/advanced
- Will not repull all internal transactions as this will probably
affect too few people and transactions and repulling them all would
affect all users. The solution for affected users is to go and repull
the affected transactions
- Adjusted validator requests decoding so it works fine for multiple
consolidation even if the fees are missing and to work fine for
multiple requests. Added a test.