Fix the issue with premium tests encrypted DB
remote_encrypted_db.bin in tests contained `combined_trades_view`
which is deleted in v34->v35. But the DB is mistakenly at v35. The
existence of that combined trades view breaks the swaps removal in
v47->v48 so we fix it here.

The fix is a hacky mock instead of actually fixing the file because
... I thought it faster as I would have to figure out how to recreate
the file and I don't have that time.