Try to checksum address when adding them into the addressbook
This is done since when the blockchain is marked as None there isn't any
check but this can lead to evm addresses that are not checksummed being
added to the book. Those are useless since when queried they are usually queried
with the blockchain and checksummed in the validation before querying the database
making it miss the address that we are looking for.