Fix swaps receiving EURe in gnosis where both v1/v2 received
There is two bugs here.

1. The receival was actually removed due to that special monerium
handling logic.
2. Specific to paraswap, and before removing it, it was seeing that
the user was geting two in events and was considering them of refunds
and as such reducing the actual amount of send, in some cases making
it look as if you are swapping negative numbers

Unfortunately also needed to modify the way exceptions are since at
some decoders like cowswap we check raw trades and in the raw trades
the trade log event uses the old and not new monerium. It's a mess