Fix Swaps event identifier
Should fix https://github.com/rotki/rotki/issues/10096

This was a nasty bug.

The way the SwapEvent class was initialized it was trying to create
its own event identifier if one was not passed.

The issue is that SwapEvent does not represent an entire swap, but
only a single side. Spend, receive or fee. Which meant that if no
history event identifier was passed each one of them was creating its
own event identifier making 2/3 events invisible and breaking the
swaps.

What this commit does is move this functionality outside of SwapEvent
and into the responsibility of the caller of `create_swap_events`.

The SwapEvent class design, just like the AssetMovement class design
is very brittle and duplicating everything. We should change it.

The commit is huge due to needing to touch virtually all swap related tests.